+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -855,7 +855,7 @@ export interface Complexity {
 	/**
 	 * Disallow unnecessary catch clauses.
 	 */
-	noUselessCatch?: RuleConfiguration_for_Null;
+	noUselessCatch?: RuleFixConfiguration_for_Null;
 	/**
 	 * Disallow unnecessary constructors.
 	 */